+++ lib/reio.dart
@@ -1,6 +1,6 @@
-export 'src/component.dart';
-export 'src/observable_stores.dart';
-export 'src/virtual_node.dart';
-export 'src/virtual_node_alias.dart';
-export 'src/virtual_node_controller.dart';
-export 'src/virtual_node_observable.dart';
+// Component
+export 'src/component/wrapper.dart';
+export 'src/component/construction/wrapper.dart';
+
+// Tools
+export 'src/tools.dart';