+++ test/web/main.dart
@@ -1,10 +1,13 @@
-import 'components/app.reio.dart';
+import 'dart:html';
+
+import 'components/app.dart';
+import 'components/test.dart';
 
 void main() {
   // If this is your first time using Reio, it is highly
   // it is highly recommended that you open the documentation of the framework.
   // https://github.com/MineEjo/reiodart/wiki
 
-  /// Creating a component App.
-  $App();
+  app.inject(document.getElementById('app'));
+  test.mount(app);
 }