+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+## 0.2.0
+
+* Added tools.
+* Fixed the critical bug of inserting children of a node.
+* Fixed bug of ignoring new attributes.
+* Fixed a critical bug on store subscription.
+* Added components.
+* Added automation and helpers for components.
+* Code refactoring and optimization.
+
 ## 0.1.4
 
 * Important decisions and tests have been made.